Move all analysis functions to log_recovery_analysis files
