+++ cubrid/CMakeLists.txt
@@ -308,6 +308,7 @@ set(TRANSACTION_SOURCES
   ${TRANSACTION_DIR}/log_prior_recv.cpp
   ${TRANSACTION_DIR}/log_reader.cpp
   ${TRANSACTION_DIR}/log_recovery.c
+  ${TRANSACTION_DIR}/log_recovery_analysis.cpp
   ${TRANSACTION_DIR}/log_recovery_context.cpp
   ${TRANSACTION_DIR}/log_recovery_redo.cpp
   ${TRANSACTION_DIR}/log_recovery_redo_parallel.cpp
@@ -342,6 +343,7 @@ set(TRANSACTION_HEADERS
   ${TRANSACTION_DIR}/log_reader.hpp
   ${TRANSACTION_DIR}/log_record.hpp
   ${TRANSACTION_DIR}/log_recovery.h
+  ${TRANSACTION_DIR}/log_recovery_analysis.hpp
   ${TRANSACTION_DIR}/log_recovery_context.hpp
   ${TRANSACTION_DIR}/log_recovery_redo.hpp
   ${TRANSACTION_DIR}/log_recovery_redo_parallel.hpp
